Rearrange modules and little refactoring
