+++ prover/src/storage_proof/storage_trie_proof/generic_hasher.rs
