+++ prover/src/storage_proof/storage_trie_proof/hashed_branch_parser.rs
