+++ prover/src/storage_proof/storage_trie_proof/hashed_leaf_parser.rs
