+++ prover/src/storage_proof/storage_trie_proof/node_parser/branch_parser/child_node_array_parser/scale_compact_integer_parser.rs
